+++ source/agModule.cs
@@ -126,8 +126,8 @@ namespace RoverAntiGravitySystem
 
             isOn = false;
 
-            this.gameObject.GetChild("Box002").renderer.material.SetColor("_EmissiveColor", Color.black);
-            this.gameObject.GetChild("Box003").renderer.material.SetColor("_EmissiveColor", Color.red);
+            this.gameObject.GetChild("buttonGreen").renderer.material.SetColor("_EmissiveColor", Color.black);
+            this.gameObject.GetChild("buttonRed").renderer.material.SetColor("_EmissiveColor", Color.red);
         }
 
         public override void OnActive()
@@ -165,25 +165,31 @@ namespace RoverAntiGravitySystem
         {
             base.OnUpdate();
 
+            if (Input.GetKeyDown(KeyCode.Keypad5))
+            {
+                isOn = !isOn;
+                isSwitching = true;
+            }
+
             if (isSwitching && isOn)
             {
                 this.myAnimation["switch"].time = 0;
-                this.myAnimation["switch"].speed = 5;
+                this.myAnimation["switch"].speed = 1;
                 this.myAnimation.Play("switch");
 
-                this.gameObject.GetChild("Box002").renderer.material.SetColor("_EmissiveColor", Color.green);
-                this.gameObject.GetChild("Box003").renderer.material.SetColor("_EmissiveColor", Color.black);
+                this.gameObject.GetChild("buttonGreen").renderer.material.SetColor("_EmissiveColor", Color.green);
+                this.gameObject.GetChild("buttonRed").renderer.material.SetColor("_EmissiveColor", Color.black);
 
                 isSwitching = false;
             }
             else if (isSwitching && !isOn)
             {
                 this.myAnimation["switch"].time = this.myAnimation["switch"].length;
-                this.myAnimation["switch"].speed = -5;
+                this.myAnimation["switch"].speed = -1;
                 this.myAnimation.Play("switch");
 
-                this.gameObject.GetChild("Box002").renderer.material.SetColor("_EmissiveColor", Color.black);
-                this.gameObject.GetChild("Box003").renderer.material.SetColor("_EmissiveColor", Color.red);
+                this.gameObject.GetChild("buttonGreen").renderer.material.SetColor("_EmissiveColor", Color.black);
+                this.gameObject.GetChild("buttonRed").renderer.material.SetColor("_EmissiveColor", Color.red);
 
                 isSwitching = false;
             }
@@ -215,8 +221,8 @@ namespace RoverAntiGravitySystem
             {
                 this.myAnimation["switch"].time = 0;
 
-                this.gameObject.GetChild("Box002").renderer.material.SetColor("_EmissiveColor", Color.black);
-                this.gameObject.GetChild("Box003").renderer.material.SetColor("_EmissiveColor", Color.red);
+                this.gameObject.GetChild("buttonGreen").renderer.material.SetColor("_EmissiveColor", Color.black);
+                this.gameObject.GetChild("buttonRed").renderer.material.SetColor("_EmissiveColor", Color.red);
             }
         }